+++ packages/mock-addon-docs/stories/Button.stories.jsx
@@ -5,20 +5,20 @@ import { mockData } from './mock';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Example/Button',
-  component: Button,
-  // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
-  argTypes: {
-    backgroundColor: { control: 'color' },
-  },
-  parameters: {
-    myAddonParameter: `
+    title: 'Example/Button',
+    component: Button,
+    // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
+    argTypes: {
+        backgroundColor: { control: 'color' },
+    },
+    parameters: {
+        myAddonParameter: `
 <MyComponent boolProp scalarProp={1} complexProp={{ foo: 1, bar: '2' }}>
   <SomeOtherComponent funcProp={(a) => a.id} />
 </MyComponent>
 `,
-    mockData: ['global'],
-  },
+        mockData: [],
+    },
 };
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
@@ -27,29 +27,29 @@ const Template = (args) => <Button {...args} />;
 export const Primary = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 Primary.args = {
-  primary: true,
-  label: 'Button',
+    primary: true,
+    label: 'Button',
 };
 Primary.parameters = {
-  mockData,
-}
+    mockData,
+};
 
 export const Secondary = Template.bind({});
 Secondary.args = {
-  label: 'Button',
+    label: 'Button',
 };
 Secondary.parameters = {
-  mockData: ['hello'],
-}
+    mockData: ['hello'],
+};
 
 export const Large = Template.bind({});
 Large.args = {
-  size: 'large',
-  label: 'Button',
+    size: 'large',
+    label: 'Button',
 };
 
 export const Small = Template.bind({});
 Small.args = {
-  size: 'small',
-  label: 'Button',
+    size: 'small',
+    label: 'Button',
 };
