+++ packages/mock-addon-docs/stories/Page.stories.jsx
@@ -1,25 +1,25 @@
-import React from 'react';
-import { within, userEvent } from '@storybook/testing-library';
+// import React from 'react';
+// import { within, userEvent } from '@storybook/testing-library';
 
-import { Page } from './Page';
+// import { Page } from './Page';
 
-export default {
-  title: 'Example/Page',
-  component: Page,
-  parameters: {
-    // More on Story layout: https://storybook.js.org/docs/react/configure/story-layout
-    layout: 'fullscreen',
-  },
-};
+// export default {
+//   title: 'Example/Page',
+//   component: Page,
+//   parameters: {
+//     // More on Story layout: https://storybook.js.org/docs/react/configure/story-layout
+//     layout: 'fullscreen',
+//   },
+// };
 
-const Template = (args) => <Page {...args} />;
+// const Template = (args) => <Page {...args} />;
 
-// More on interaction testing: https://storybook.js.org/docs/react/writing-tests/interaction-testing
-export const LoggedOut = Template.bind({});
+// // More on interaction testing: https://storybook.js.org/docs/react/writing-tests/interaction-testing
+// export const LoggedOut = Template.bind({});
 
-export const LoggedIn = Template.bind({});
-LoggedIn.play = async ({ canvasElement }) => {
-  const canvas = within(canvasElement);
-  const loginButton = await canvas.getByRole('button', { name: /Log in/i });
-  await userEvent.click(loginButton);
-};
+// export const LoggedIn = Template.bind({});
+// LoggedIn.play = async ({ canvasElement }) => {
+//   const canvas = within(canvasElement);
+//   const loginButton = await canvas.getByRole('button', { name: /Log in/i });
+//   await userEvent.click(loginButton);
+// };
