+++ packages/mock-addon/src/Panel.js
@@ -1,31 +1,17 @@
-import React, { useEffect } from 'react';
-import { useAddonState, useChannel, useParameter } from '@storybook/api';
+import React from 'react';
+import { useAddonState, useChannel } from '@storybook/api';
 import { AddonPanel, Placeholder, ScrollArea } from '@storybook/components';
 
-import { ADDON_ID, EVENTS, PARAM_KEY } from './utils/constants';
-import faker from './utils/faker';
+import { ADDON_ID, EVENTS } from './utils/constants';
 
 import { MockItem } from './components/MockItem';
 
 export const Panel = (props) => {
     const [mockData, setState] = useAddonState(ADDON_ID, []);
 
-    const paramData = useParameter(PARAM_KEY, []);
-
     const emit = useChannel({
-        [EVENTS.SEND]: (newMockData) => {
-            setState(newMockData);
-        },
-        [EVENTS.UPDATE]: (item, name, value) => {
-            faker.update(item, name, value);
-            emit(EVENTS.SEND, faker.getRequests());
-        },
+        [EVENTS.SEND]: (newMockData) => setState(newMockData),
     });
-
-    useEffect(() => {
-        faker.makeInitialRequestMap(paramData);
-        emit(EVENTS.SEND, faker.getRequests());
-    }, [paramData]);
 
     const onChange = (item, name, value) => {
         emit(EVENTS.UPDATE, item, name, value);
