+++ packages/mock-addon/src/preset/preview.js
@@ -1,0 +1,3 @@
+import { withRoundTrip } from '../withRoundTrip';
+
+export const decorators = [withRoundTrip];
